Add default value in retrieving environment variables and replace request.get_data with request.data
In this commit, we added default values in retrieving environment variables to prevent errors caused by None when the variable is not found. We also replaced request.get_data with request.data.decode("utf-8") in the app.route("/callback", methods=["POST"]) method to ensure that the received data is encoded in UTF-8. Additionally, we created a dictionary for the required environment variables to reduce code redundancy.